feat: adapted fetch to work on serverSide
