+++ apps/web/middleware.ts
@@ -1,0 +1,12 @@
+import { NextResponse } from 'next/server'
+
+export function middleware(request) {
+  const requestHeaders = new Headers(request.headers)
+  requestHeaders.set('x-pathname', request.nextUrl.pathname)
+
+  return NextResponse.next({
+    request: {
+      headers: requestHeaders,
+    },
+  })
+}
